+++ bastion/resource_authdomain_ldap.go
@@ -6,68 +6,64 @@ import (
 	"fmt"
 	"net/http"
 
-	govers "github.com/hashicorp/go-version"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
 	bchk "github.com/jeremmfr/go-utils/basiccheck"
 )
 
-type jsonLdapDomain struct {
+type jsonAuthDomainLdap struct {
 	CheckX509SanEmail    bool     `json:"check_x509_san_email"`
 	IsDefault            bool     `json:"is_default"`
-	DomainName           string   `json:"domain_name,omitempty"`
-	DefaultLanguage      string   `json:"default_language"`
+	ID                   string   `json:"id,omitempty"`
+	AuthDomainName       string   `json:"auth_domain_name"`
 	DefaultEmailDomain   string   `json:"default_email_domain"`
+	DefaultLanguage      string   `json:"default_language"`
 	Description          string   `json:"description"`
 	DisplayNameAttribute string   `json:"display_name_attribute"`
+	DomainName           string   `json:"domain_name"`
 	EmailAttribute       string   `json:"email_attribute"`
-	LdapDomainName       string   `json:"ldap_domain_name"`
-	LanguageAttribute    string   `json:"language_attribute"`
 	GroupAttribute       string   `json:"group_attribute"`
+	LanguageAttribute    string   `json:"language_attribute"`
+	PubKeyAttribute      string   `json:"pubkey_attribute"`
 	SanDomainName        string   `json:"san_domain_name"`
+	Type                 string   `json:"type"`
 	X509Condition        string   `json:"x509_condition"`
 	X509SearchFilter     string   `json:"x509_search_filter"`
-	ExternalLdaps        []string `json:"external_ldaps"`
+	ExternalAuths        []string `json:"external_auths"`
 	SecondaryAuth        []string `json:"secondary_auth"`
 }
 
-func resourceLdapDomain() *schema.Resource {
+func resourceAuthDomainLdap() *schema.Resource {
 	return &schema.Resource{
-		CreateContext: resourceLdapDomainCreate,
-		ReadContext:   resourceLdapDomainRead,
-		UpdateContext: resourceLdapDomainUpdate,
-		DeleteContext: resourceLdapDomainDelete,
+		CreateContext: resourceAuthDomainLdapCreate,
+		ReadContext:   resourceAuthDomainLdapRead,
+		UpdateContext: resourceAuthDomainLdapUpdate,
+		DeleteContext: resourceAuthDomainLdapDelete,
 		Importer: &schema.ResourceImporter{
-			State: resourceLdapDomainImport,
+			State: resourceAuthDomainLdapImport,
 		},
 		Schema: map[string]*schema.Schema{
 			"domain_name": {
 				Type:     schema.TypeString,
 				Required: true,
-				ForceNew: true,
 			},
-			"ldap_domain_name": {
+			"auth_domain_name": {
 				Type:     schema.TypeString,
 				Required: true,
 			},
-			"external_ldaps": {
-				Type:     schema.TypeList,
+			"default_email_domain": {
+				Type:     schema.TypeString,
 				Required: true,
-				Elem:     &schema.Schema{Type: schema.TypeString},
 			},
 			"default_language": {
 				Type:         schema.TypeString,
 				Required:     true,
 				ValidateFunc: validation.StringInSlice([]string{"de", "en", "es", "fr", "ru"}, false),
 			},
-			"default_email_domain": {
-				Type:     schema.TypeString,
-				Required: true,
-			},
-			"secondary_auth": {
+			"external_auths": {
 				Type:     schema.TypeList,
-				Optional: true,
+				Required: true,
 				Elem:     &schema.Schema{Type: schema.TypeString},
 			},
 			"description": {
@@ -98,9 +94,18 @@ func resourceLdapDomain() *schema.Resource {
 				Type:     schema.TypeString,
 				Optional: true,
 			},
+			"pubkey_attribute": {
+				Type:     schema.TypeString,
+				Optional: true,
+			},
 			"san_domain_name": {
 				Type:     schema.TypeString,
 				Optional: true,
+			},
+			"secondary_auth": {
+				Type:     schema.TypeList,
+				Optional: true,
+				Elem:     &schema.Schema{Type: schema.TypeString},
 			},
 			"x509_condition": {
 				Type:     schema.TypeString,
@@ -113,128 +118,155 @@ func resourceLdapDomain() *schema.Resource {
 		},
 	}
 }
-func resourceLdapDomainVersionCheck(version string) error {
-	if bchk.InSlice(version, []string{VersionWallixAPI33, VersionWallixAPI36}) {
+
+func resourceAuthDomainLdapVersionCheck(version string) error {
+	if bchk.InSlice(version, []string{VersionWallixAPI38}) {
 		return nil
 	}
-	if vers, err := govers.NewVersion(version); err == nil {
-		versionResourceRename, _ := govers.NewVersion(VersionWallixAPI38)
-		if vers.GreaterThanOrEqual(versionResourceRename) {
-			return fmt.Errorf(
-				"resource wallix-bastion_ldapdomain not available with api version %s\n"+
-					" use wallix-bastion_authdomain_ldap instead",
-				version)
-		}
-	}
 
-	return fmt.Errorf("resource wallix-bastion_ldapdomain not available with api version %s", version)
+	return fmt.Errorf("resource wallix-bastion_authdomain_ldap not available with api version %s", version)
 }
 
-func resourceLdapDomainCreate(ctx context.Context, d *schema.ResourceData, m interface{}) diag.Diagnostics {
+func resourceAuthDomainLdapCreate(
+	ctx context.Context, d *schema.ResourceData, m interface{},
+) diag.Diagnostics {
 	c := m.(*Client)
-	if err := resourceLdapDomainVersionCheck(c.bastionAPIVersion); err != nil {
+	if err := resourceAuthDomainLdapVersionCheck(c.bastionAPIVersion); err != nil {
 		return diag.FromErr(err)
 	}
-	ex, err := checkResourceLdapDomainExists(ctx, d.Get("domain_name").(string), m)
+	_, ex, err := searchResourceAuthDomainLdap(ctx, d.Get("domain_name").(string), m)
 	if err != nil {
 		return diag.FromErr(err)
 	}
 	if ex {
 		return diag.FromErr(fmt.Errorf("domain_name %s already exists", d.Get("domain_name").(string)))
 	}
-	err = addLdapDomain(ctx, d, m)
+	err = addAuthDomainLdap(ctx, d, m)
 	if err != nil {
 		return diag.FromErr(err)
 	}
-	d.SetId(d.Get("domain_name").(string))
+	id, ex, err := searchResourceAuthDomainLdap(ctx, d.Get("domain_name").(string), m)
+	if err != nil {
+		return diag.FromErr(err)
+	}
+	if !ex {
+		return diag.FromErr(fmt.Errorf("domain_name %s not found after POST", d.Get("domain_name").(string)))
+	}
+	d.SetId(id)
 
-	return resourceLdapDomainRead(ctx, d, m)
+	return resourceAuthDomainLdapRead(ctx, d, m)
 }
-func resourceLdapDomainRead(ctx context.Context, d *schema.ResourceData, m interface{}) diag.Diagnostics {
+
+func resourceAuthDomainLdapRead(
+	ctx context.Context, d *schema.ResourceData, m interface{},
+) diag.Diagnostics {
 	c := m.(*Client)
-	if err := resourceLdapDomainVersionCheck(c.bastionAPIVersion); err != nil {
+	if err := resourceAuthDomainLdapVersionCheck(c.bastionAPIVersion); err != nil {
 		return diag.FromErr(err)
 	}
-	cfg, err := readLdapDomainOptions(ctx, d.Id(), m)
+	cfg, err := readAuthDomainLdapOptions(ctx, d.Id(), m)
 	if err != nil {
 		return diag.FromErr(err)
 	}
-	if cfg.DomainName == "" {
+	if cfg.ID == "" {
 		d.SetId("")
 	} else {
-		fillLdapDomain(d, cfg)
+		fillAuthDomainLdap(d, cfg)
 	}
 
 	return nil
 }
-func resourceLdapDomainUpdate(ctx context.Context, d *schema.ResourceData, m interface{}) diag.Diagnostics {
+
+func resourceAuthDomainLdapUpdate(
+	ctx context.Context, d *schema.ResourceData, m interface{},
+) diag.Diagnostics {
 	d.Partial(true)
 	c := m.(*Client)
-	if err := resourceLdapDomainVersionCheck(c.bastionAPIVersion); err != nil {
+	if err := resourceAuthDomainLdapVersionCheck(c.bastionAPIVersion); err != nil {
 		return diag.FromErr(err)
 	}
-	if err := updateLdapDomain(ctx, d, m); err != nil {
+	if err := updateAuthDomainLdap(ctx, d, m); err != nil {
 		return diag.FromErr(err)
 	}
 	d.Partial(false)
 
-	return resourceLdapDomainRead(ctx, d, m)
+	return resourceAuthDomainLdapRead(ctx, d, m)
 }
-func resourceLdapDomainDelete(ctx context.Context, d *schema.ResourceData, m interface{}) diag.Diagnostics {
+
+func resourceAuthDomainLdapDelete(
+	ctx context.Context, d *schema.ResourceData, m interface{},
+) diag.Diagnostics {
 	c := m.(*Client)
-	if err := resourceLdapDomainVersionCheck(c.bastionAPIVersion); err != nil {
+	if err := resourceAuthDomainLdapVersionCheck(c.bastionAPIVersion); err != nil {
 		return diag.FromErr(err)
 	}
-	if err := deleteLdapDomain(ctx, d, m); err != nil {
+	if err := deleteAuthDomainLdap(ctx, d, m); err != nil {
 		return diag.FromErr(err)
 	}
 
 	return nil
 }
-func resourceLdapDomainImport(d *schema.ResourceData, m interface{}) ([]*schema.ResourceData, error) {
+
+func resourceAuthDomainLdapImport(
+	d *schema.ResourceData, m interface{},
+) (
+	[]*schema.ResourceData, error,
+) {
 	ctx := context.Background()
 	c := m.(*Client)
-	if err := resourceLdapDomainVersionCheck(c.bastionAPIVersion); err != nil {
+	if err := resourceAuthDomainLdapVersionCheck(c.bastionAPIVersion); err != nil {
 		return nil, err
 	}
-	ex, err := checkResourceLdapDomainExists(ctx, d.Id(), m)
+	id, ex, err := searchResourceAuthDomainLdap(ctx, d.Id(), m)
 	if err != nil {
 		return nil, err
 	}
 	if !ex {
 		return nil, fmt.Errorf("don't find domain_name with id %s (id must be <domain_name>", d.Id())
 	}
-	cfg, err := readLdapDomainOptions(ctx, d.Id(), m)
+	cfg, err := readAuthDomainLdapOptions(ctx, d.Id(), m)
 	if err != nil {
 		return nil, err
 	}
-	fillLdapDomain(d, cfg)
+	fillAuthDomainLdap(d, cfg)
 	result := make([]*schema.ResourceData, 1)
+	d.SetId(id)
 	result[0] = d
 
 	return result, nil
 }
 
-func checkResourceLdapDomainExists(ctx context.Context, domainName string, m interface{}) (bool, error) {
+func searchResourceAuthDomainLdap(
+	ctx context.Context, domainName string, m interface{},
+) (
+	string, bool, error,
+) {
 	c := m.(*Client)
-	body, code, err := c.newRequest(ctx, "/ldapdomains/"+domainName, http.MethodGet, nil)
+	body, code, err := c.newRequest(ctx, "/authdomains/?q=domain_name="+domainName, http.MethodGet, nil)
 	if err != nil {
-		return false, err
-	}
-	if code == http.StatusNotFound {
-		return false, nil
+		return "", false, err
 	}
 	if code != http.StatusOK {
-		return false, fmt.Errorf("api doesn't return OK : %d with body :\n%s", code, body)
+		return "", false, fmt.Errorf("api doesn't return OK : %d with body :\n%s", code, body)
+	}
+	var results []jsonAuthDomainLdap
+	err = json.Unmarshal([]byte(body), &results)
+	if err != nil {
+		return "", false, fmt.Errorf("json.Unmarshal failed : %w", err)
+	}
+	if len(results) == 1 {
+		return results[0].ID, true, nil
 	}
 
-	return true, nil
+	return "", false, nil
 }
 
-func addLdapDomain(ctx context.Context, d *schema.ResourceData, m interface{}) error {
+func addAuthDomainLdap(
+	ctx context.Context, d *schema.ResourceData, m interface{},
+) error {
 	c := m.(*Client)
-	jsonData := prepareLdapDomainJSON(d, true)
-	body, code, err := c.newRequest(ctx, "/ldapdomains/", http.MethodPost, jsonData)
+	jsonData := prepareAuthDomainLdapJSON(d)
+	body, code, err := c.newRequest(ctx, "/authdomains/", http.MethodPost, jsonData)
 	if err != nil {
 		return err
 	}
@@ -245,10 +277,12 @@ func addLdapDomain(ctx context.Context, d *schema.ResourceData, m interface{}) e
 	return nil
 }
 
-func updateLdapDomain(ctx context.Context, d *schema.ResourceData, m interface{}) error {
+func updateAuthDomainLdap(
+	ctx context.Context, d *schema.ResourceData, m interface{},
+) error {
 	c := m.(*Client)
-	jsonData := prepareLdapDomainJSON(d, false)
-	body, code, err := c.newRequest(ctx, "/ldapdomains/"+d.Id()+"?force=true", http.MethodPut, jsonData)
+	jsonData := prepareAuthDomainLdapJSON(d)
+	body, code, err := c.newRequest(ctx, "/authdomains/"+d.Id()+"?force=true", http.MethodPut, jsonData)
 	if err != nil {
 		return err
 	}
@@ -259,9 +293,11 @@ func updateLdapDomain(ctx context.Context, d *schema.ResourceData, m interface{}
 	return nil
 }
 
-func deleteLdapDomain(ctx context.Context, d *schema.ResourceData, m interface{}) error {
+func deleteAuthDomainLdap(
+	ctx context.Context, d *schema.ResourceData, m interface{},
+) error {
 	c := m.(*Client)
-	body, code, err := c.newRequest(ctx, "/ldapdomains/"+d.Id(), http.MethodDelete, nil)
+	body, code, err := c.newRequest(ctx, "/authdomains/"+d.Id(), http.MethodDelete, nil)
 	if err != nil {
 		return err
 	}
@@ -272,11 +308,14 @@ func deleteLdapDomain(ctx context.Context, d *schema.ResourceData, m interface{}
 	return nil
 }
 
-func prepareLdapDomainJSON(d *schema.ResourceData, newResource bool) jsonLdapDomain {
-	jsonData := jsonLdapDomain{
-		LdapDomainName:       d.Get("ldap_domain_name").(string),
-		DefaultLanguage:      d.Get("default_language").(string),
+func prepareAuthDomainLdapJSON(d *schema.ResourceData) jsonAuthDomainLdap {
+	jsonData := jsonAuthDomainLdap{
+		Type:                 "LDAP",
+		DomainName:           d.Get("domain_name").(string),
+		AuthDomainName:       d.Get("auth_domain_name").(string),
 		DefaultEmailDomain:   d.Get("default_email_domain").(string),
+		DefaultLanguage:      d.Get("default_language").(string),
+		ExternalAuths:        make([]string, 0),
 		Description:          d.Get("description").(string),
 		CheckX509SanEmail:    d.Get("check_x509_san_email").(bool),
 		DisplayNameAttribute: d.Get("display_name_attribute").(string),
@@ -284,17 +323,15 @@ func prepareLdapDomainJSON(d *schema.ResourceData, newResource bool) jsonLdapDom
 		GroupAttribute:       d.Get("group_attribute").(string),
 		IsDefault:            d.Get("is_default").(bool),
 		LanguageAttribute:    d.Get("language_attribute").(string),
+		PubKeyAttribute:      d.Get("pubkey_attribute").(string),
 		SanDomainName:        d.Get("san_domain_name").(string),
+		SecondaryAuth:        make([]string, 0),
 		X509Condition:        d.Get("x509_condition").(string),
 		X509SearchFilter:     d.Get("x509_search_filter").(string),
 	}
-	if newResource {
-		jsonData.DomainName = d.Get("domain_name").(string)
+	for _, v := range d.Get("external_auths").([]interface{}) {
+		jsonData.ExternalAuths = append(jsonData.ExternalAuths, v.(string))
 	}
-	for _, v := range d.Get("external_ldaps").([]interface{}) {
-		jsonData.ExternalLdaps = append(jsonData.ExternalLdaps, v.(string))
-	}
-	jsonData.SecondaryAuth = make([]string, 0)
 	for _, v := range d.Get("secondary_auth").([]interface{}) {
 		jsonData.SecondaryAuth = append(jsonData.SecondaryAuth, v.(string))
 	}
@@ -302,11 +339,14 @@ func prepareLdapDomainJSON(d *schema.ResourceData, newResource bool) jsonLdapDom
 	return jsonData
 }
 
-func readLdapDomainOptions(
-	ctx context.Context, domainName string, m interface{}) (jsonLdapDomain, error) {
+func readAuthDomainLdapOptions(
+	ctx context.Context, domainID string, m interface{},
+) (
+	jsonAuthDomainLdap, error,
+) {
 	c := m.(*Client)
-	var result jsonLdapDomain
-	body, code, err := c.newRequest(ctx, "/ldapdomains/"+domainName, http.MethodGet, nil)
+	var result jsonAuthDomainLdap
+	body, code, err := c.newRequest(ctx, "/authdomains/"+domainID, http.MethodGet, nil)
 	if err != nil {
 		return result, err
 	}
@@ -324,14 +364,14 @@ func readLdapDomainOptions(
 	return result, nil
 }
 
-func fillLdapDomain(d *schema.ResourceData, jsonData jsonLdapDomain) {
+func fillAuthDomainLdap(d *schema.ResourceData, jsonData jsonAuthDomainLdap) {
 	if tfErr := d.Set("domain_name", jsonData.DomainName); tfErr != nil {
 		panic(tfErr)
 	}
-	if tfErr := d.Set("ldap_domain_name", jsonData.LdapDomainName); tfErr != nil {
+	if tfErr := d.Set("auth_domain_name", jsonData.AuthDomainName); tfErr != nil {
 		panic(tfErr)
 	}
-	if tfErr := d.Set("external_ldaps", jsonData.ExternalLdaps); tfErr != nil {
+	if tfErr := d.Set("external_auths", jsonData.ExternalAuths); tfErr != nil {
 		panic(tfErr)
 	}
 	if tfErr := d.Set("default_language", jsonData.DefaultLanguage); tfErr != nil {
@@ -359,6 +399,9 @@ func fillLdapDomain(d *schema.ResourceData, jsonData jsonLdapDomain) {
 		panic(tfErr)
 	}
 	if tfErr := d.Set("language_attribute", jsonData.LanguageAttribute); tfErr != nil {
+		panic(tfErr)
+	}
+	if tfErr := d.Set("pubkey_attribute", jsonData.PubKeyAttribute); tfErr != nil {
 		panic(tfErr)
 	}
 	if tfErr := d.Set("san_domain_name", jsonData.SanDomainName); tfErr != nil {