+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 FEATURES:
 
 * add `wallix-bastion_externalauth_saml` resource
+* add `wallix-bastion_authdomain_ldap` resource
 
 ENHANCEMENTS: